refactor(Navigation): renaming AppNavigation to Navigation, and AppNavigation/Authentication to AuthNavigation
