+++ src/modules/App/App.tsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
 import { persistor, store } from 'redux/store'
 
-import AppNavigation from '../AppNavigation'
+import AppNavigation from '../Navigation/Navigation.container'
 
 setI18nConfig()
