+++ src/modules/Navigation/Navigation.tsx
@@ -2,8 +2,8 @@ import { NavigationContainer } from '@react-navigation/native'
 import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 
+import AuthNavigation from '../AuthNavigation/AuthNavigation'
 import Home from '../Home/Home'
-import Authentication from './Authentication/Authentication'
 import linking from './Linking'
 
 const Stack = createStackNavigator()
@@ -12,16 +12,16 @@ type Props = {
   isAuthorised: boolean
 }
 
-const AppNavigation = ({ isAuthorised }: Props) => (
+const Navigation = ({ isAuthorised }: Props) => (
   <NavigationContainer linking={linking}>
     <Stack.Navigator headerMode="none">
       {isAuthorised ? (
         <Stack.Screen name="Home" component={Home} />
       ) : (
-        <Stack.Screen name="Authentication" component={Authentication} />
+        <Stack.Screen name="Authentication" component={AuthNavigation} />
       )}
     </Stack.Navigator>
   </NavigationContainer>
 )
 
-export default AppNavigation
+export default Navigation
