+++ src/modules/Navigation/index.ts
@@ -1,0 +1,3 @@
+import AppNavigation from './Navigation.container'
+
+export default AppNavigation
